# rust-ssr-example-frontend/lib/__generated__/models/dateFilterModel.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * rust_ssr_example_backend
 * OpenAPI spec version: 0.1.0
 */

export interface DateFilterModel {
  /** @nullable */
  dateFrom?: string | null;
  /** @nullable */
  dateTo?: string | null;
  type: string;
}