# rust-ssr-example-frontend/lib/__generated__/models/filterModel.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * rust_ssr_example_backend
 * OpenAPI spec version: 0.1.0
 */
import type { DateFilterModel } from './dateFilterModel';
import type { NumberFilterModel } from './numberFilterModel';
import type { TextFilterModel } from './textFilterModel';

export type FilterModel = TextFilterModel & {
  filterType: 'text';
} | NumberFilterModel & {
  filterType: 'number';
} | DateFilterModel & {
  filterType: 'date';
};