# rust-ssr-example-frontend/lib/__generated__/models/getRowsRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * rust_ssr_example_backend
 * OpenAPI spec version: 0.1.0
 */
import type { GetRowsRequestFilterModel } from './getRowsRequestFilterModel';
import type { SortModelItem } from './sortModelItem';

export interface GetRowsRequest {
  endRow: number;
  filterModel?: GetRowsRequestFilterModel;
  sortModel?: SortModelItem[];
  startRow: number;
}