# rust-ssr-example-frontend/lib/__generated__/models/numberFilterModel.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * rust_ssr_example_backend
 * OpenAPI spec version: 0.1.0
 */

export interface NumberFilterModel {
  /** @nullable */
  filter?: number | null;
  /** @nullable */
  filterTo?: number | null;
  type: string;
}