# rust-ssr-example-frontend/lib/__generated__/models/stockHistoryGetRowsResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * rust_ssr_example_backend
 * OpenAPI spec version: 0.1.0
 */
import type { StockHistoryRow } from './stockHistoryRow';

export interface StockHistoryGetRowsResponse {
  rowCount: number;
  rowData: StockHistoryRow[];
}