# rust-ssr-example-frontend/lib/__generated__/models/stockHistoryRow.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * rust_ssr_example_backend
 * OpenAPI spec version: 0.1.0
 */

export interface StockHistoryRow {
  close: string;
  date: string;
  high: string;
  low: string;
  open: string;
  ticker: string;
  /** @minimum 0 */
  volume: number;
}