# rust-ssr-example-frontend/lib/__generated__/models/textFilterModel.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * rust_ssr_example_backend
 * OpenAPI spec version: 0.1.0
 */

export interface TextFilterModel {
  /** @nullable */
  filter?: string | null;
  type: string;
}