# rust_ssr_example_backend/openapi.yml
openapi: 3.0.3
info:
  title: rust_ssr_example_backend
  description: ''
  license:
    name: ''
  version: 0.1.0
paths:
  /api/stock-history:
    get:
      tags:
      - domains::stock_history::routes
      operationId: list_stock_history
      responses:
        '200':
          description: List stock history rows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StockHistoryRow'
        '500':
          description: Database error
  /api/stock-history/getRows:
    post:
      tags:
      - domains::stock_history::routes
      summary: 'AG Grid Server-Side Row Model datasource endpoint: infinite scroll, sort'
      description: and filter. See https://www.ag-grid.com/react-data-grid/server-side-model-datasource/
      operationId: get_rows
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetRowsRequest'
        required: true
      responses:
        '200':
          description: Server-side rows for AG Grid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockHistoryGetRowsResponse'
        '400':
          description: Invalid request
        '500':
          description: Database error
components:
  schemas:
    DateFilterModel:
      type: object
      required:
      - type
      properties:
        dateFrom:
          type: string
          nullable: true
        dateTo:
          type: string
          nullable: true
        type:
          type: string
    FilterModel:
      oneOf:
      - allOf:
        - $ref: '#/components/schemas/TextFilterModel'
        - type: object
          required:
          - filterType
          properties:
            filterType:
              type: string
              enum:
              - text
      - allOf:
        - $ref: '#/components/schemas/NumberFilterModel'
        - type: object
          required:
          - filterType
          properties:
            filterType:
              type: string
              enum:
              - number
      - allOf:
        - $ref: '#/components/schemas/DateFilterModel'
        - type: object
          required:
          - filterType
          properties:
            filterType:
              type: string
              enum:
              - date
      discriminator:
        propertyName: filterType
    GetRowsRequest:
      type: object
      required:
      - startRow
      - endRow
      properties:
        endRow:
          type: integer
          format: int64
        filterModel:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/FilterModel'
        sortModel:
          type: array
          items:
            $ref: '#/components/schemas/SortModelItem'
        startRow:
          type: integer
          format: int64
    NumberFilterModel:
      type: object
      required:
      - type
      properties:
        filter:
          type: number
          format: double
          nullable: true
        filterTo:
          type: number
          format: double
          nullable: true
        type:
          type: string
    SortModelItem:
      type: object
      required:
      - colId
      - sort
      properties:
        colId:
          type: string
        sort:
          type: string
    StockHistoryGetRowsResponse:
      type: object
      required:
      - rowData
      - rowCount
      properties:
        rowCount:
          type: integer
          format: int64
        rowData:
          type: array
          items:
            $ref: '#/components/schemas/StockHistoryRow'
    StockHistoryRow:
      type: object
      required:
      - date
      - ticker
      - open
      - high
      - low
      - close
      - volume
      properties:
        close:
          type: string
        date:
          type: string
          format: date
        high:
          type: string
        low:
          type: string
        open:
          type: string
        ticker:
          type: string
        volume:
          type: integer
          format: int64
          minimum: 0
    TextFilterModel:
      type: object
      required:
      - type
      properties:
        filter:
          type: string
          nullable: true
        type:
          type: string